Create 2/3 cases for delete method
Create leaf? method on bst.rb to know if a node is a leafe
Create one_child? method on bst.rb to know ifa node has only one one_child
Create one_child method that returns the only child of that node
Create antecessor method on Tree class, that gets the parent of the input node
Create leaf case for delete method. If the node is a leaf, it just gets deleted
Create ony child case for delete method. If the node has only one child,
the node's child replace the node.